+++ packages/class-core/src/deltas/inverters.ts
@@ -26,10 +26,10 @@ import {
     AnnotationReplacedDelta,
     ChildAddedDelta,
     ChildDeletedDelta,
-    ChildMovedAndReplacedFromOtherContainmentDelta,
+    ChildMovedAndReplacedFromContainmentInOtherParentDelta,
     ChildMovedAndReplacedFromOtherContainmentInSameParentDelta,
     ChildMovedAndReplacedInSameContainmentDelta,
-    ChildMovedFromOtherContainmentDelta,
+    ChildMovedFromContainmentInOtherParentDelta,
     ChildMovedFromOtherContainmentInSameParentDelta,
     ChildMovedInSameContainmentDelta,
     ChildReplacedDelta,
@@ -76,8 +76,8 @@ export const invertDelta = (delta: IDelta): IDelta => {
     if (delta instanceof ChildReplacedDelta) {
         return new ChildReplacedDelta(delta.parent, delta.containment, delta.index, delta.newChild, delta.replacedChild);
     }
-    if (delta instanceof ChildMovedFromOtherContainmentDelta) {
-        return new ChildMovedFromOtherContainmentDelta(delta.newParent, delta.newContainment, delta.newIndex, delta.oldParent, delta.oldContainment, delta.oldIndex, delta.movedChild);
+    if (delta instanceof ChildMovedFromContainmentInOtherParentDelta) {
+        return new ChildMovedFromContainmentInOtherParentDelta(delta.newParent, delta.newContainment, delta.newIndex, delta.oldParent, delta.oldContainment, delta.oldIndex, delta.movedChild);
     }
     if (delta instanceof ChildMovedFromOtherContainmentInSameParentDelta) {
         return new ChildMovedFromOtherContainmentInSameParentDelta(delta.parent, delta.newContainment, delta.newIndex, delta.movedChild, delta.oldContainment, delta.oldIndex);
@@ -86,9 +86,9 @@ export const invertDelta = (delta: IDelta): IDelta => {
         const [invertedOldIndex, invertedIndexOffset] = invertedMoveWithOffset(delta.oldIndex, delta.indexOffset);
         return new ChildMovedInSameContainmentDelta(delta.parent, delta.containment, invertedOldIndex, invertedIndexOffset, delta.movedChild);
     }
-    if (delta instanceof ChildMovedAndReplacedFromOtherContainmentDelta) {
+    if (delta instanceof ChildMovedAndReplacedFromContainmentInOtherParentDelta) {
         return new CompositeDelta([
-            new ChildMovedFromOtherContainmentDelta(delta.newParent, delta.newContainment, delta.newIndex, delta.oldParent, delta.oldContainment, delta.oldIndex, delta.movedChild),
+            new ChildMovedFromContainmentInOtherParentDelta(delta.newParent, delta.newContainment, delta.newIndex, delta.oldParent, delta.oldContainment, delta.oldIndex, delta.movedChild),
             new ChildAddedDelta(delta.newParent, delta.newContainment, delta.newIndex, delta.replacedChild)
         ]);
     }