+++ packages/class-core/src/deltas/serialization/deserializer.g.ts
@@ -34,10 +34,10 @@ import {
     AnnotationReplacedDelta,
     ChildAddedDelta,
     ChildDeletedDelta,
-    ChildMovedAndReplacedFromOtherContainmentDelta,
+    ChildMovedAndReplacedFromContainmentInOtherParentDelta,
     ChildMovedAndReplacedFromOtherContainmentInSameParentDelta,
     ChildMovedAndReplacedInSameContainmentDelta,
-    ChildMovedFromOtherContainmentDelta,
+    ChildMovedFromContainmentInOtherParentDelta,
     ChildMovedFromOtherContainmentInSameParentDelta,
     ChildMovedInSameContainmentDelta,
     ChildReplacedDelta,
@@ -107,7 +107,7 @@ export const deltaDeserializer = (languageBases: ILanguageBase[], idMapping: IdM
                 const newChild = idMapping.nodeBaseFromId(delta.newChild);
                 return new ChildReplacedDelta(parent, containment, index, replacedChild, newChild);
             }
-            case "ChildMovedFromOtherContainment": {
+            case "ChildMovedFromContainmentInOtherParent": {
                 const oldParent = idMapping.nodeBaseFromId(delta.oldParent);
                 const oldContainment = resolvedContainmentFrom(delta.oldContainment, oldParent.classifier);
                 const oldIndex = delta.oldIndex;
@@ -115,7 +115,7 @@ export const deltaDeserializer = (languageBases: ILanguageBase[], idMapping: IdM
                 const newContainment = resolvedContainmentFrom(delta.newContainment, newParent.classifier);
                 const newIndex = delta.newIndex;
                 const movedChild = idMapping.nodeBaseFromId(delta.movedChild);
-                return new ChildMovedFromOtherContainmentDelta(oldParent, oldContainment, oldIndex, newParent, newContainment, newIndex, movedChild);
+                return new ChildMovedFromContainmentInOtherParentDelta(oldParent, oldContainment, oldIndex, newParent, newContainment, newIndex, movedChild);
             }
             case "ChildMovedFromOtherContainmentInSameParent": {
                 const parent = idMapping.nodeBaseFromId(delta.parent);
@@ -134,7 +134,7 @@ export const deltaDeserializer = (languageBases: ILanguageBase[], idMapping: IdM
                 const movedChild = idMapping.nodeBaseFromId(delta.movedChild);
                 return new ChildMovedInSameContainmentDelta(parent, containment, oldIndex, indexOffset, movedChild);
             }
-            case "ChildMovedAndReplacedFromOtherContainment": {
+            case "ChildMovedAndReplacedFromContainmentInOtherParent": {
                 const newParent = idMapping.nodeBaseFromId(delta.newParent);
                 const newContainment = resolvedContainmentFrom(delta.newContainment, newParent.classifier);
                 const newIndex = delta.newIndex;
@@ -143,7 +143,7 @@ export const deltaDeserializer = (languageBases: ILanguageBase[], idMapping: IdM
                 const oldContainment = resolvedContainmentFrom(delta.oldContainment, oldParent.classifier);
                 const oldIndex = delta.oldIndex;
                 const replacedChild = idMapping.nodeBaseFromId(delta.replacedChild);
-                return new ChildMovedAndReplacedFromOtherContainmentDelta(newParent, newContainment, newIndex, movedChild, oldParent, oldContainment, oldIndex, replacedChild);
+                return new ChildMovedAndReplacedFromContainmentInOtherParentDelta(newParent, newContainment, newIndex, movedChild, oldParent, oldContainment, oldIndex, replacedChild);
             }
             case "ChildMovedAndReplacedFromOtherContainmentInSameParent": {
                 const parent = idMapping.nodeBaseFromId(delta.parent);