+++ packages/delta-protocol-common/src/translators/delta-to-command.ts
@@ -25,10 +25,10 @@ import {
     AnnotationReplacedDelta,
     ChildAddedDelta,
     ChildDeletedDelta,
-    ChildMovedAndReplacedFromOtherContainmentDelta,
+    ChildMovedAndReplacedFromContainmentInOtherParentDelta,
     ChildMovedAndReplacedFromOtherContainmentInSameParentDelta,
     ChildMovedAndReplacedInSameContainmentDelta,
-    ChildMovedFromOtherContainmentDelta,
+    ChildMovedFromContainmentInOtherParentDelta,
     ChildMovedFromOtherContainmentInSameParentDelta,
     ChildMovedInSameContainmentDelta,
     ChildReplacedDelta,
@@ -167,8 +167,8 @@ export const deltaToCommandTranslator = (
                 replacedChild: delta.replacedChild.id
             })
         }
-        if (delta instanceof ChildMovedFromOtherContainmentDelta) {
-            return completed<MoveChildFromOtherContainmentCommand>("MoveChildFromOtherContainment", { // § 5.7.5.4
+        if (delta instanceof ChildMovedFromContainmentInOtherParentDelta) {
+            return completed<MoveChildFromOtherContainmentCommand>("MoveChildFromContainmentInOtherParent", { // § 5.7.5.4
                 newParent: delta.newParent.id,
                 newContainment: metaPointerForFeature(delta.newContainment),
                 newIndex: delta.newIndex,
@@ -197,8 +197,8 @@ export const deltaToCommandTranslator = (
                 movedChild: delta.movedChild.id
             })
         }
-        if (delta instanceof ChildMovedAndReplacedFromOtherContainmentDelta) {
-            return completed<MoveAndReplaceChildFromOtherContainmentCommand>("MoveAndReplaceChildFromOtherContainment", { // § 5.7.5.7
+        if (delta instanceof ChildMovedAndReplacedFromContainmentInOtherParentDelta) {
+            return completed<MoveAndReplaceChildFromOtherContainmentCommand>("MoveAndReplaceChildFromContainmentInOtherParent", { // § 5.7.5.7
                 newParent: delta.newParent.id,
                 newContainment: metaPointerForFeature(delta.newContainment),
                 newIndex: delta.newIndex,