+++ packages/delta-protocol-repository-ws/src/command-to-event.ts
@@ -183,9 +183,9 @@ export const commandAsEvent = (command: Command, participationId: string): Event
                     index
                 })
             }
-            case "MoveChildFromOtherContainment": {
+            case "MoveChildFromContainmentInOtherParent": {
                 const {newParent, newContainment, newIndex, movedChild, oldParent, oldContainment, oldIndex} = command as MoveChildFromOtherContainmentCommand // § 5.7.5.4
-                return completed<ChildMovedFromOtherContainmentEvent>("ChildMovedFromOtherContainment", { // § 5.8.5.4
+                return completed<ChildMovedFromOtherContainmentEvent>("ChildMovedFromContainmentInOtherParent", { // § 5.8.5.4
                     newParent,
                     newContainment,
                     newIndex,
@@ -216,9 +216,9 @@ export const commandAsEvent = (command: Command, participationId: string): Event
                     oldIndex
                 })
             }
-            case "MoveAndReplaceChildFromOtherContainment": {
+            case "MoveAndReplaceChildFromContainmentInOtherParent": {
                 const {newParent, newContainment, newIndex, replacedChild, movedChild, oldParent, oldContainment, oldIndex} = command as MoveAndReplaceChildFromOtherContainmentCommand // § 5.7.5.7
-                return completed<ChildMovedAndReplacedFromOtherContainmentEvent>("ChildMovedAndReplacedFromOtherContainment", { // § 5.8.5.7
+                return completed<ChildMovedAndReplacedFromOtherContainmentEvent>("ChildMovedAndReplacedFromContainmentInOtherParent", { // § 5.8.5.7
                     newParent,
                     newContainment,
                     newIndex,