+++ packages/delta-protocol-repository-ws/src/repository-impl.ts
@@ -99,10 +99,10 @@ export class LionWebRepository {
                 case "AddChild":
                 case "DeleteChild":
                 case "ReplaceChild":
-                case "MoveChildFromOtherContainment":
+                case "MoveChildFromContainmentInOtherParent":
                 case "MoveChildFromOtherContainmentInSameParent":
                 case "MoveChildInSameContainment":
-                case "MoveAndReplaceChildFromOtherContainment":
+                case "MoveAndReplaceChildFromContainmentInOtherParent":
                 case "MoveAndReplaceChildFromOtherContainmentInSameParent":
                 case "MoveAndReplaceChildInSameContainment":
                 case "AddAnnotation":